Initialize CVS branch matt-nb5-mips64
The per-file CVS branch roots do not exactly match a reconstructed source-branch commit.

+++ usr.sbin/rtadvd/advcap.h
@@ -1,4 +1,5 @@
-/*	$NetBSD: advcap.h,v 1.2 1999/07/06 13:02:09 itojun Exp $	*/
+/*	$NetBSD: advcap.h,v 1.6 2006/03/05 23:47:08 rpaulo Exp $	*/
+/*	$KAME: advcap.h,v 1.5 2003/06/09 05:40:54 t-momose Exp $	*/
 
 /*
  * Copyright (C) 1994,1995 by Andrey A. Chernov, Moscow, Russia.
@@ -37,7 +38,7 @@ __BEGIN_DECLS
 
 extern int agetent __P((char *, const char *));
 extern int agetflag __P((const char *));
-extern int agetnum __P((const char *));
+extern int64_t agetnum __P((const char *));
 extern char *agetstr __P((const char *, char **));
 
 __END_DECLS
